+++ pandera/typing/common.py
@@ -9,10 +9,6 @@ import typing_inspect
 
 from pandera import dtypes
 from pandera.engines import numpy_engine, pandas_engine
-from pandera.typing import pyspark_sql
-
-if pyspark_sql.PYSPARK_SQL_INSTALLED:
-    from pandera.engines import pyspark_engine
 
 Bool = dtypes.Bool  #: ``"bool"`` numpy dtype
 Date = dtypes.Date  #: ``datetime.date`` object dtype
@@ -47,18 +43,6 @@ String = dtypes.String  #: ``"str"`` numpy dtype
 #: fall back on the str-as-object-array representation.
 STRING = pandas_engine.STRING  #: ``"str"`` numpy dtype
 BOOL = pandas_engine.BOOL  #: ``"str"`` numpy dtype
-if pyspark_sql.PYSPARK_SQL_INSTALLED:
-    PYSPARK_STRING = pyspark_engine.String
-    PYSPARK_INT = pyspark_engine.Int
-    PYSPARK_LONGINT = pyspark_engine.BigInt
-    PYSPARK_SHORTINT = pyspark_engine.ShortInt
-    PYSPARK_BYTEINT = pyspark_engine.ByteInt
-    PYSPARK_DOUBLE = pyspark_engine.Double
-    PYSPARK_FLOAT = pyspark_engine.Float
-    PYSPARK_DECIMAL = pyspark_engine.Decimal
-    PYSPARK_DATE = pyspark_engine.Date
-    PYSPARK_TIMESTAMP = pyspark_engine.Timestamp
-    PYSPARK_BINARY = pyspark_engine.Binary
 
 try:
     Geometry = pandas_engine.Geometry  # : ``"geometry"`` geopandas dtype
@@ -150,22 +134,6 @@ else:
         ],
     )
 
-if pyspark_sql.PYSPARK_SQL_INSTALLED:
-    PysparkDType = TypeVar(  # type: ignore
-        "GenericDtype",
-        bound=Union[
-            PYSPARK_STRING,
-            PYSPARK_INT,
-            PYSPARK_LONGINT,
-            PYSPARK_SHORTINT,
-            PYSPARK_BYTEINT,
-            PYSPARK_FLOAT,
-            PYSPARK_DECIMAL,
-            PYSPARK_DATE,
-            PYSPARK_TIMESTAMP,
-            PYSPARK_BINARY,
-        ],
-    )
 
 DataFrameModel = TypeVar("Schema", bound="DataFrameModel")  # type: ignore
 
@@ -231,20 +199,6 @@ class IndexBase(Generic[GenericDtype]):
 
     def __get__(self, instance: object, owner: Type) -> str:  # pragma: no cover
         raise AttributeError("Indexes should resolve to pa.Index-s")
-
-
-if pyspark_sql.PYSPARK_SQL_INSTALLED:
-
-    class ColumnBase(Generic[PysparkDType]):
-        """Representation of pandas.Index, only used for type annotation.
-
-        *new in 0.5.0*
-        """
-
-        default_dtype: Optional[Type] = None
-
-        def __get__(self, instance: object, owner: Type) -> str:  # pragma: no cover
-            raise AttributeError("column should resolve to pyspark.sql.Column-s")
 
 
 class AnnotationInfo:  # pylint:disable=too-few-public-methods
