+++ pandera/typing/pyspark_sql.py
@@ -1,4 +1,7 @@
 """Pandera type annotations for Dask."""
+from pandera.typing.common import DataFrameBase, GenericDtype
+from pandera.typing.pandas import DataFrameModel, _GenericAlias
+from typing import Union, Optional, Type, TypeVar
 
 try:
     import pyspark.sql as ps
@@ -8,10 +11,37 @@ except ImportError:  # pragma: no cover
     PYSPARK_SQL_INSTALLED = False
 
 if PYSPARK_SQL_INSTALLED:
-    from typing import TYPE_CHECKING, Generic, TypeVar
+    from pandera.engines import pyspark_engine
 
-    from pandera.typing.common import ColumnBase, DataFrameBase, GenericDtype
-    from pandera.typing.pandas import DataFrameModel, _GenericAlias
+    PYSPARK_STRING = pyspark_engine.String
+    PYSPARK_INT = pyspark_engine.Int
+    PYSPARK_LONGINT = pyspark_engine.BigInt
+    PYSPARK_SHORTINT = pyspark_engine.ShortInt
+    PYSPARK_BYTEINT = pyspark_engine.ByteInt
+    PYSPARK_DOUBLE = pyspark_engine.Double
+    PYSPARK_FLOAT = pyspark_engine.Float
+    PYSPARK_DECIMAL = pyspark_engine.Decimal
+    PYSPARK_DATE = pyspark_engine.Date
+    PYSPARK_TIMESTAMP = pyspark_engine.Timestamp
+    PYSPARK_BINARY = pyspark_engine.Binary
+
+
+    PysparkDType = TypeVar(  # type: ignore
+        "PysparkDType",
+        bound=Union[
+                    PYSPARK_STRING,
+                    PYSPARK_INT,
+                    PYSPARK_LONGINT,
+                    PYSPARK_SHORTINT,
+                    PYSPARK_BYTEINT,
+                    PYSPARK_FLOAT,
+                    PYSPARK_DECIMAL,
+                    PYSPARK_DATE,
+                    PYSPARK_TIMESTAMP,
+                    PYSPARK_BINARY,
+        ],
+    )
+    from typing import TYPE_CHECKING, Generic, TypeVar
 
     # pylint:disable=invalid-name
     if TYPE_CHECKING:
@@ -21,6 +51,18 @@ if PYSPARK_SQL_INSTALLED:
 
     if PYSPARK_SQL_INSTALLED:
         # pylint: disable=too-few-public-methods,arguments-renamed
+        class ColumnBase(Generic[PysparkDType]):
+            """Representation of pandas.Index, only used for type annotation.
+
+            *new in 0.5.0*
+            """
+
+            default_dtype: Optional[Type] = None
+
+            def __get__(self, instance: object, owner: Type) -> str:  # pragma: no cover
+                raise AttributeError("column should resolve to pyspark.sql.Column-s")
+
+
         class DataFrame(DataFrameBase, ps.DataFrame, Generic[T]):
             """
             Representation of dask.dataframe.DataFrame, only used for type
